move page components into page scoped directories
