+++ src/pages/create-tenant/CreateTenant.tsx
@@ -11,8 +11,8 @@ import type { Metadata } from '@/types/tenants'
 import { toast } from 'sonner'
 import ErrorDisplay from '@/components/ErrorDisplay'
 import Button from '@/components/Button'
-import ContactInformation from '../components/ContactInformation'
-import InfrastructureMetadata from '../components/InfrastructureMetadata'
+import ContactInformation from './ContactInformation'
+import InfrastructureMetadata from './InfrastructureMetadata'
 import LoadingSpinner from '@/components/LoadingSpinner'
 import PageHeader from '@/components/PageHeader'
 import Tabs from '@/components/Tabs'
