+++ src/pages/tenant-details/TenantDetails.tsx
